// discovery/plugins/notifications/src/types/notifications.ts
export type AppEmailNotification = {
  receiver_user_id: number
} & NotificationRow
// ...
export type USDCPurchaseBuyerNotification = {
  content_type: string
  buyer_user_id: number
  seller_user_id: number
  amount: number
  content_id: number
}